+++ packages/reactive-core/src/reaction.ts
@@ -1,11 +1,10 @@
-import { $reactive, ObserverConnection } from "./observable";
 import { Observer } from "./observer";
 
 let runningReactions: Reaction[] = [];
-
-export class Reaction implements Observer {
-  public observing = new Set<ObserverConnection<object>>();
-  constructor(private func: () => void | Promise<void>, private options: { name: string }) {}
+export class Reaction extends Observer {
+  constructor(private func: () => void | Promise<void>, private options: { name: string }) {
+    super(() => this._trigger());
+  }
 
   private reaction = () => {
     runningReactions.push(this);
@@ -17,20 +16,12 @@ export class Reaction implements Observer {
     }
   };
 
-  public trigger() {
+  private _trigger() {
     // TODO: catch errors
     if (runningReactions.includes(this)) {
       throw new Error("already running reaction");
     }
-    this.observing.forEach((val) => {
-      const connections = val.source.observable[$reactive].connections;
-      if (val.source.type === "iterate") {
-        connections.iterate.delete(this);
-      } else {
-        connections.byKey.get(val.source.key).delete(this);
-      }
-    });
-    this.observing.clear();
+    this.removeObservers();
     this.reaction();
   }
 }
